requirements.txt
fastapi
uvicorn
pydantic
sqlite3
PyJWT
bcrypt
streamlit
requests
beautifulsoup4
pytz
python-dotenv
os 
sys
re 
logging 
datetime




